Extract test utilities
