+++ misk-jobqueue/src/test/kotlin/misk/jobqueue/testutilities/Color.kt
@@ -1,0 +1,8 @@
+package misk.jobqueue.testutilities
+
+internal enum class Color {
+  RED,
+  GREEN
+}
+
+internal class ColorException : Throwable()
