+++ misk-jobqueue/src/test/kotlin/misk/jobqueue/testutilities/ExampleJob.kt
@@ -1,0 +1,7 @@
+package misk.jobqueue.testutilities
+
+internal data class ExampleJob(
+  val color: Color,
+  val message: String,
+  val hint: ExampleJobHint? = null
+)
