+++ misk-jobqueue/src/test/kotlin/misk/jobqueue/testutilities/ExampleJobHint.kt
@@ -1,0 +1,9 @@
+package misk.jobqueue.testutilities
+
+internal enum class ExampleJobHint {
+  DONT_ACK,
+  THROW,
+  THROW_ONCE,
+  DEAD_LETTER,
+  DEAD_LETTER_ONCE
+}
