+++ misk-jobqueue/src/test/kotlin/misk/jobqueue/testutilities/Queues.kt
@@ -1,0 +1,7 @@
+package misk.jobqueue.testutilities
+
+import misk.jobqueue.QueueName
+
+internal val RED_QUEUE = QueueName("red_queue")
+internal val GREEN_QUEUE = QueueName("green_queue")
+internal val ENQUEUER_QUEUE = QueueName("first_step_queue")
